svm-sdk-tests passes again.
