+++ crates/sdk-tests/tests/meta/ctor_meta.rs
@@ -9,7 +9,7 @@ mod Template {
 }
 
 fn main() {
-    let raw = raw_schema();
+    let raw = raw_meta();
     let json: Value = serde_json::from_str(&raw).unwrap();
 
     assert_eq!(
