+++ crates/sdk-tests/tests/meta/endpoint_with_params_array_meta.rs
@@ -1,21 +1,16 @@
+#![allow(unused)]
 use serde_json::{json, Value};
 
-use svm_sdk::template;
+use svm_sdk::{template, Address, Amount};
 
 #[template]
 mod Template {
-    #[fundable]
     #[endpoint]
-    fn call() {}
-
-    #[fundable_hook(default)]
-    fn fund() {
-        //
-    }
+    fn call(a: [bool; 3]) {}
 }
 
 fn main() {
-    let raw = raw_schema();
+    let raw = raw_meta();
     let json: Value = serde_json::from_str(&raw).unwrap();
 
     assert_eq!(
@@ -26,9 +21,11 @@ fn main() {
                 "api_name": "call",
                 "wasm_name": "call",
                 "is_ctor": false,
-                "is_fundable": true,
+                "is_fundable": false,
                 "doc": "",
-                "signature": json!({"params": [], "returns": {}}),
+                "signature": json!({"params": [
+                    json!({"name": "a", "type": "[bool]", "length": 3}),
+                ], "returns": {}}),
             })],
         })
     );
