+++ crates/sdk-tests/tests/meta/endpoint_with_params_meta.rs
@@ -10,7 +10,7 @@ mod Template {
 }
 
 fn main() {
-    let raw = raw_schema();
+    let raw = raw_meta();
     let json: Value = serde_json::from_str(&raw).unwrap();
 
     assert_eq!(
