+++ crates/sdk-tests/tests/meta/storage_meta.rs
@@ -24,7 +24,7 @@ mod Template {
 }
 
 fn main() {
-    let raw = raw_schema();
+    let raw = raw_meta();
     let json: Value = serde_json::from_str(&raw).unwrap();
 
     assert_eq!(
